requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
docx2txt==0.9
fastapi==0.116.1
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
Mako==1.3.10
MarkupSafe==3.0.2
nodeenv==1.9.1
openai==1.98.0
orjson==3.11.1
packaging==25.0
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.23.0
